[Model Refactoring] Move deepseek_v4_ops to models/deepseek_v4
