+++ vllm/models/deepseek_v4/common/ops/__init__.py
@@ -8,7 +8,6 @@ from .cache_utils import (
     quantize_and_insert_k_cache,
 )
 from .fused_indexer_q import MXFP4_BLOCK_SIZE, fused_indexer_q_rope_quant
-from .fused_inv_rope_fp8_quant import fused_inv_rope_fp8_quant
 from .fused_qk_rmsnorm import fused_q_kv_rmsnorm
 
 __all__ = [
@@ -17,7 +16,6 @@ __all__ = [
     "compute_global_topk_indices_and_lens",
     "dequantize_and_gather_k_cache",
     "fused_indexer_q_rope_quant",
-    "fused_inv_rope_fp8_quant",
     "fused_q_kv_rmsnorm",
     "quantize_and_insert_k_cache",
 ]
