+++ vllm/models/deepseek_v4/common/ops/fused_compress_quant_cache.py
